requirements.txt
pandas
numpy
scikit-learn
pyyaml
transformers
langdetect
nltk
tqdm
mlflow
torch
seaborn
matplotlib
spacy